- implementata libreria FEA e verifica - integrata funzionalità FEA in CIEID - fix issue #32

+++ cie-pkcs11/Util/UUCStringTable.cpp
@@ -88,7 +88,7 @@ unsigned long UUCStringTable::getHash(const char* szKey)
 
 	if (len < 16) 
 	{
- 	    for (unsigned long i = len ; i > 0; i--) 
+ 	    for (int i = len ; i > 0; i--)
 		{
  			h = (h * 37) + val[off++];
  	    }
@@ -97,7 +97,7 @@ unsigned long UUCStringTable::getHash(const char* szKey)
 	{
  	    // only sample some characters
  	    unsigned long skip = len / 8;
- 	    for (unsigned long i = len ; i > 0; i -= skip, off += skip)
+ 	    for (int i = len ; i > 0; i -= skip, off += skip)
 		{
  			h = (h * 39) + val[off];
  	    }
